dom/html/test/test_non-ascii-cookie.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=784367
-->
<head>
  <meta charset="utf-8">
  <title>Test for non-ASCII cookie values</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=784367">Mozilla Bug 784367</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for non-ASCII cookie values **/

var [Cc, Ci] = [SpecialPowers.Cc, SpecialPowers.Ci];

var cm = Cc["@mozilla.org/cookiemanager;1"].getService(Ci.nsICookieManager);
function getCookieFromManager() {
  var values = [];
  var host = location.hostname;
  var path = location.pathname;
  path = path.substring(0, path.lastIndexOf("/") + 1);
  var e = cm.enumerator;
  while (e.hasMoreElements()) {
    var cookie = e.getNext().QueryInterface(Ci.nsICookie);
    if (!cookie) {
      break;
    }
    if (host != cookie.host || path != cookie.path) {
      continue;
    }
    values.push(cookie.name + "=" + cookie.value);
  }
  return values.join("; ");
}

var c = document.cookie;
is(document.cookie, 'abc=012©ABC\ufffdDEF', "document.cookie should be decoded as UTF-8");
is(getCookieFromManager(), document.cookie, "nsICookieManager should be consistent with document.cookie");
var newCookie = 'def=∼≩≭≧∯≳≲≣∽≸≸∺≸∠≯≮≥≲≲≯≲∽≡≬≥≲≴∨∱∩∾';
document.cookie = newCookie;
is(document.cookie, c + '; ' + newCookie, "document.cookie should be encoded as UTF-8");
is(getCookieFromManager(), document.cookie, "nsICookieManager should be consistent with document.cookie");
var date1 = new Date();
date1.setTime(0);
document.cookie = newCookie + 'def=;expires=' + date1.toGMTString();
</script>
</pre>
</body>
</html>
